BrainTumor/requirements.txt
ultralytics
fastapi
Pillow
uvicorn
opencv-python
numpy
python-multipart